Create array target sum triplet.cpp
Input Format:-

1. First line contains input N.
2. Next line contains N space separated integers denoting the elements of the array.
3. The third line contains a single integer T denoting the target element.

Output Format:-
Print all the triplet present in the array in a new line each. The triplets must be printed as A, B and C where A,B and C are the elements of the triplet ( A<=B<=C) and all triplets must be printed in sorted order. Print only unique triplets.

Input
9
5 7 9 1 2 4 6 8 3
10

Output
1, 2 and 7
1, 3 and 6
1, 4 and 5
2, 3 and 5