Create maximum subarray.cpp
The first line consists of number of test cases T. Each test case consists of two lines.
The first line of each testcase contains a single integer N denoting the number of elements for the array A.
The second line of each testcase contains N space separated integers denoting the elements of the array.

We need to find contigious elements whose sum is greater.

Input
2
4
1 2 3 4
3
-10 5 10

Output
10
15